+++ ModCore/Api/Entities/ApiInvite.cs
@@ -1,6 +1,6 @@
 ﻿namespace ModCore.Api.Entities
 {
-    public class ApiInvite
+    public struct ApiInvite
     {
         public string Invite { get; set; }
     }
