+++ ModCore/Commands/Note.cs
@@ -10,6 +10,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using System.Linq;
+using ModCore.Database.Entities;
 
 namespace ModCore.Commands
 {
