+++ ModCore/Commands/Reminders.cs
@@ -10,6 +10,7 @@ using DSharpPlus.Interactivity.EventHandling;
 using Humanizer;
 using Humanizer.Localisation;
 using ModCore.Database;
+using ModCore.Database.Entities;
 using ModCore.Entities;
 using ModCore.Listeners;
 using ModCore.Logic;
