+++ ModCore/Database/Entities/DatabaseCommandId.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database
+namespace ModCore.Database.Entities
 {
     [Table("mcore_cmd_state")]
     public class DatabaseCommandId
@@ -16,7 +16,7 @@ namespace ModCore.Database
         [Annotation("Npgsql:ValueGeneratedOnAdd", true)]
         [Annotation("Sqlite:Autoincrement", true)]
         public short Id { get; set; }
-        
+
         [Key]
         [Column("command_qualified")]
         public string Command { get; set; }
