+++ ModCore/Database/Entities/DatabaseLevel.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database
+namespace ModCore.Database.Entities
 {
     [Table("mcore_levels")]
     public class DatabaseLevel
