+++ ModCore/Database/Entities/DatabaseTag.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database
+namespace ModCore.Database.Entities
 {
     [Table("mcore_tags")]
     public class DatabaseTag
@@ -16,21 +16,21 @@ namespace ModCore.Database
         [Index("guild_id_channel_id_name_key", IsUnique = true, IsLocal = true)]
         [Column("channel_id")]
         public long ChannelId { get; set; }
-        
+
         [Index("guild_id_channel_id_name_key", IsUnique = true, IsLocal = true)]
         [Column("tagname")]
         public string Name { get; set; }
         #endregion
-        
+
         [Column("id")]
         public int Id { get; set; }
-        
+
         [Column("owner_id")]
         public long OwnerId { get; set; }
-        
+
         [Column("created_at", TypeName = "timestamptz")]
         public DateTime CreatedAt { get; set; }
-        
+
         [Column("contents")]
         public string Contents { get; set; }
     }
