+++ ModCore/Entities/SharedData.cs
@@ -6,7 +6,6 @@ using System.Threading;
 using DSharpPlus;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.Entities;
-using ModCore.CoreApi;
 using ModCore.Listeners;
 
 namespace ModCore.Entities
@@ -17,13 +16,11 @@ namespace ModCore.Entities
         public DateTime ProcessStartTime { get; internal set; }
         public SemaphoreSlim TimerSempahore { get; internal set; }
         public TimerData TimerData { get; internal set; }
-        public List<ulong> BotManagers { get; internal set; }
         public string DefaultPrefix { get; internal set; }
 		public int ReadysReceived { get; internal set; } = 0;
 		public List<Permissions> AllPermissions { get; internal set; } = new List<Permissions>();
         public ConcurrentDictionary<ulong, DiscordMessage> DeletedMessages = new ConcurrentDictionary<ulong, DiscordMessage>();
         public ConcurrentDictionary<ulong, DiscordMessage> EditedMessages = new ConcurrentDictionary<ulong, DiscordMessage>();
-        public ModCoreEmojis Emojis;
 
         /// <summary>
         /// Every command, top-level or not, along with full qualified name.
@@ -42,7 +39,6 @@ namespace ModCore.Entities
         public void Initialize(ModCoreShard shard)
         {
             Commands = shard.Commands.RegisteredCommands.SelectMany(SelectCommandsFromDict).Distinct().ToArray();
-            Emojis = ModCoreEmojis.LoadEmojis(shard.Client);
         }
 
         private static IEnumerable<(string name, Command cmd)> SelectCommandsFromDict(KeyValuePair<string, Command> c)
