--- ModCore/emojis.json
@@ -1,1 +1,0 @@
-{"jumplink":577900935913930792}
